\name{pacode}
\alias{pacode}
\docType{package}
\title{Basic analytical functionality for parallelcoordinates}
\description{
  This package provides some analytical functionality for parallelcoordinates.de.
  More specifically, it provides clustering and dimensionality reduction.
}
\details{
  \tabular{ll}{
    Package: \tab pacode\cr
    Type: \tab Package\cr
    Version: \tab 0.1\cr
    Date: \tab 2015-05-29\cr
    License: \tab MIT\cr
  }
}
\author{
  Maintainer: Bertjan Broeksema <bertjan.broeksema@list.lu>
}
\references{
  https://julianheinrich.github.io/parallelcoordinates.de/paco/index.html
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%~~ documentation directory ~~
\keyword{ package, parallel coordinates, visual analytics }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
